Update while loop in gmm fit method.
Add an abs value calculation in the gmm.fit() while loop to ensure that it only terminates if the difference between log-likelihoods is small. Otherwise, it can terminate before convergence with high negative values. With the absolute values, we ensure convergence, since both log-likelihoods have to have very similar values with a difference below the threshold. This caused my tests to randomly fail.